BSAPP-929.2_Views-3, BSAPP-1049 created: LigapasseDO, LigamatchDO, LigapasseComponent, LigapasseComponentImpl, LigamatchComponent, LigamatchComponentImpl created: new file structure in Ligapasse and Ligamatch deleted: MatchPasseViewComponent.java, MatchPasseViewComponentImpl.java, MatchPasseViewDO.java
